+++ includes/Plugin.php
@@ -85,6 +85,7 @@ class Plugin extends ServiceBasedPlugin {
 		'integrations.cfi'             => Integrations\Conditional_Featured_Image::class,
 		'integrations.sitekit'         => Integrations\Site_Kit::class,
 		'integrations.themes_support'  => Integrations\Core_Themes_Support::class,
+		'integrations.shortpixel'      => Integrations\ShortPixel::class,
 		'imgareaselect_patch'          => Admin\ImgAreaSelect_Patch::class,
 		'kses'                         => KSES::class,
 		'font_post_type'               => Font_Post_Type::class,
@@ -133,7 +134,6 @@ class Plugin extends ServiceBasedPlugin {
 		'taxonomy.tag'                 => Taxonomy\Tag_Taxonomy::class,
 		'user_preferences'             => User\Preferences::class,
 		'web_stories_block'            => Block\Web_Stories_Block::class,
-		'shortpixel_patch'             => Admin\ShortPixel_Patch::class,
 	];
 
 	/**
